# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_wilfridterry_contact-list_internal_domain.Contact:
    properties:
      address:
        type: string
      author:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  github_com_wilfridterry_contact-list_internal_domain.SaveInputContact:
    properties:
      address:
        type: string
      author:
        type: string
      email:
        type: string
      last_name:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - address
    - author
    - email
    - last_name
    - name
    - phone
    type: object
  github_com_wilfridterry_contact-list_internal_domain.SignInInput:
    properties:
      email:
        maxLength: 255
        minLength: 4
        type: string
      password:
        maxLength: 70
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_wilfridterry_contact-list_internal_domain.SignUpInput:
    properties:
      email:
        maxLength: 255
        minLength: 4
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      password:
        maxLength: 70
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Contacts API
  version: "1.0"
paths:
  /auth/sign-in:
    get:
      consumes:
      - application/json
      description: refresh tokens for auth
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: refresh tokens
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: sign up with data
      parameters:
      - description: user sign up
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_wilfridterry_contact-list_internal_domain.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: sign up to the system
      tags:
      - auth
  /contacts:
    get:
      consumes:
      - application/json
      description: get contacts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_wilfridterry_contact-list_internal_domain.Contact'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: List contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: create a contact with data
      parameters:
      - description: Contact paylaod
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_wilfridterry_contact-list_internal_domain.SaveInputContact'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Create a contact
      tags:
      - contacts
  /contacts/{id}:
    delete:
      consumes:
      - application/json
      description: delete a contact by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Delete a contact
      tags:
      - contacts
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_wilfridterry_contact-list_internal_domain.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Show a contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: Update a contact with data by ID
      parameters:
      - description: Contact paylaod
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_wilfridterry_contact-list_internal_domain.SaveInputContact'
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_wilfridterry_contact-list_internal_domain.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Update a contact
      tags:
      - contacts
swagger: "2.0"